Add new application properties and update data conversions
The code updates introduce new properties to the application config files for starting specific job types. In addition, a new utility function is implemented for converting string values to integers. This function is then integrated in the relevant parts of the application to handle potential NumberFormatExceptions. A datatype in the `Religion` entity class is also modified for more accurate data representation.

+++ backend/libs/entities/src/main/java/bencoepp/livius/entities/religion/Religion.java
@@ -136,7 +136,7 @@ public class Religion {
     private Double nonReligiousPctAdherents;
     private Double otherReligionsPctAdherents;
     private Double sumPercentReligiousAdherents;
-    private Integer totalPopulationPercantageCheck;
+    private Double totalPopulationPercantageCheck;
     private Boolean dualReligion;
     private String dataType;
     public final static String TYPE_NONE = "none";
